Template engine: Jinja same engine everywhere
+ Add filters from extension
