+++ requirements.txt
@@ -7,8 +7,8 @@ certifi==2019.3.9
 cffi==1.12.3
 chardet==3.0.4
 cryptography==3.3.2
-docker==3.7.0
-docker-compose==1.25.0rc2
+docker==4.4.4
+docker-compose==1.26.2
 docker-pycreds==0.4.0
 dockerpty==0.4.1
 docopt==0.6.2
@@ -22,6 +22,7 @@ google-auth==1.6.3
 idna==2.7
 ipaddress==1.0.17
 Jinja2==2.11.3
+jinja2-ansible-filters==1.3.0
 jsonschema==2.6.0
 kubernetes==9.0.0
 MarkupSafe==1.1.1
