+++ tox.ini
@@ -6,7 +6,7 @@ whitelist_externals =
   /bin/bash
   /usr/bin/bash
 deps = -rrequirements-dev.txt
-commands = pytest test --junit-xml pkr-test-report-{envname}.xml --junit-prefix {envname}
+commands = pytest test --junit-xml pkr-test-report-{envname}.xml --junit-prefix {envname} {posargs}
 
 [testenv:pylint]
 deps = pylint
